+++ src/main/java/brig/concord/meta/ConcordFileMetaType.java
@@ -1,13 +1,16 @@
 package brig.concord.meta;
 
+import brig.concord.highlighting.ConcordHighlightingColors;
 import brig.concord.meta.model.*;
 import brig.concord.yaml.meta.model.YamlMetaType;
+import com.intellij.openapi.editor.colors.TextAttributesKey;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Supplier;
 
-public class ConcordFileMetaType extends ConcordMetaType {
+public class ConcordFileMetaType extends ConcordMetaType implements HighlightProvider {
 
     private static final ConcordFileMetaType INSTANCE = new ConcordFileMetaType();
 
@@ -35,5 +38,10 @@ public class ConcordFileMetaType extends ConcordMetaType {
     @Override
     protected Map<String, Supplier<YamlMetaType>> getFeatures() {
         return features;
+    }
+
+    @Override
+    public @Nullable TextAttributesKey getKeyHighlight(String key) {
+        return ConcordHighlightingColors.DSL_SECTION;
     }
 }