+++ src/main/java/brig/concord/meta/model/FlowsMetaType.java
@@ -1,5 +1,8 @@
 package brig.concord.meta.model;
 
+import brig.concord.highlighting.ConcordHighlightingColors;
+import brig.concord.meta.HighlightProvider;
+import com.intellij.openapi.editor.colors.TextAttributesKey;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import brig.concord.yaml.meta.model.Field;
@@ -8,7 +11,7 @@ import brig.concord.yaml.psi.YAMLMapping;
 
 import java.util.List;
 
-public class FlowsMetaType extends MapMetaType {
+public class FlowsMetaType extends MapMetaType implements HighlightProvider {
 
     private static final FlowsMetaType INSTANCE = new FlowsMetaType();
 
@@ -25,32 +28,15 @@ public class FlowsMetaType extends MapMetaType {
     @Override
     public @NotNull List<Field> computeKeyCompletions(@Nullable YAMLMapping existingMapping) {
         return defaultCompletions;
-//        String placeholder = "flowName";
-//        return List.of(new Field(placeholder, getMapEntryType(null)) {
-//            @Override
-//            public @NotNull List<LookupElementBuilder> getKeyLookups(@NotNull YamlMetaType ownerClass, @NotNull PsiElement insertedScalar) {
-//                LookupElementBuilder lookup = LookupElementBuilder
-//                        .create(new TypeFieldPair(ownerClass, this), getName())
-//                        .withTypeText(getMapEntryType(null).getDisplayName(), true)
-//                        .withIcon(getLookupIcon())
-//                        .withStrikeoutness(isDeprecated())
-//                        .withInsertHandler((context, item) -> {
-//                            Editor editor = context.getEditor();
-//                            editor.getCaretModel().moveToOffset(context.getStartOffset());
-//                            SelectionModel model = editor.getSelectionModel();
-//                            model.setSelection(context.getStartOffset(), context.getStartOffset() + placeholder.length());
-//                        });
-//
-//                if (isRequired()) {
-//                    lookup = lookup.bold();
-//                }
-//                return Collections.singletonList(lookup);
-//            }
-//        });
     }
 
     @Override
     protected YamlMetaType getMapEntryType(String name) {
         return StepsMetaType.getInstance();
+    }
+
+    @Override
+    public @Nullable TextAttributesKey getKeyHighlight(String key) {
+        return ConcordHighlightingColors.FLOW_IDENTIFIER;
     }
 }