+++ src/main/java/brig/concord/meta/model/LoopMetaType.java
@@ -1,14 +1,18 @@
 package brig.concord.meta.model;
 
+import brig.concord.highlighting.ConcordHighlightingColors;
 import brig.concord.meta.ConcordMetaType;
+import brig.concord.meta.HighlightProvider;
 import brig.concord.yaml.meta.model.YamlEnumType;
 import brig.concord.yaml.meta.model.YamlMetaType;
+import com.intellij.openapi.editor.colors.TextAttributesKey;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Supplier;
 
-public class LoopMetaType extends ConcordMetaType {
+public class LoopMetaType extends ConcordMetaType implements HighlightProvider  {
 
     private static final LoopMetaType INSTANCE = new LoopMetaType();
 
@@ -36,6 +40,11 @@ public class LoopMetaType extends ConcordMetaType {
     @Override
     protected Set<String> getRequiredFields() {
         return requiredFeatures;
+    }
+
+    @Override
+    public @Nullable TextAttributesKey getKeyHighlight(String key) {
+        return ConcordHighlightingColors.DSL_KEY;
     }
 
     private static class ModeType extends YamlEnumType {