+++ src/main/java/brig/concord/meta/model/ManualTriggerEntryMetaType.java
@@ -1,14 +1,18 @@
 package brig.concord.meta.model;
 
+import brig.concord.highlighting.ConcordHighlightingColors;
 import brig.concord.meta.ConcordMetaType;
+import brig.concord.meta.HighlightProvider;
 import brig.concord.meta.model.call.CallMetaType;
 import brig.concord.yaml.meta.model.YamlMetaType;
+import com.intellij.openapi.editor.colors.TextAttributesKey;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Supplier;
 
-public class ManualTriggerEntryMetaType extends ConcordMetaType {
+public class ManualTriggerEntryMetaType extends ConcordMetaType implements HighlightProvider {
 
     private static final ManualTriggerEntryMetaType INSTANCE = new ManualTriggerEntryMetaType();
 
@@ -37,5 +41,10 @@ public class ManualTriggerEntryMetaType extends ConcordMetaType {
     @Override
     protected Set<String> getRequiredFields() {
         return required;
+    }
+
+    @Override
+    public @Nullable TextAttributesKey getKeyHighlight(String key) {
+        return ConcordHighlightingColors.DSL_KEY;
     }
 }