+++ src/main/java/brig/concord/meta/model/ProfileConfigurationMetaType.java
@@ -2,18 +2,22 @@ package brig.concord.meta.model;
 
 import brig.concord.ConcordBundle;
 import brig.concord.documentation.Documented;
+import brig.concord.highlighting.ConcordHighlightingColors;
 import brig.concord.meta.ConcordMetaType;
+import brig.concord.meta.HighlightProvider;
 import brig.concord.meta.model.call.CallMetaType;
 import brig.concord.yaml.meta.model.YamlBooleanType;
 import brig.concord.yaml.meta.model.YamlEnumType;
 import brig.concord.yaml.meta.model.YamlIntegerType;
 import brig.concord.yaml.meta.model.YamlMetaType;
+import com.intellij.openapi.editor.colors.TextAttributesKey;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Supplier;
 
-public class ProfileConfigurationMetaType extends ConcordMetaType implements Documented {
+public class ProfileConfigurationMetaType extends ConcordMetaType implements Documented, HighlightProvider {
 
     private static final ProfileConfigurationMetaType INSTANCE = new ProfileConfigurationMetaType();
 
@@ -53,5 +57,10 @@ public class ProfileConfigurationMetaType extends ConcordMetaType implements Doc
     @Override
     public String getDescription() {
         return ConcordBundle.message("Configuration.description");
+    }
+
+    @Override
+    public @Nullable TextAttributesKey getKeyHighlight(String key) {
+        return ConcordHighlightingColors.DSL_KEY;
     }
 }